fix: reinsert spinner
