--- packages/eyes-storybook/src/updateTestResults.js
@@ -1,14 +1,0 @@
-function didTestPass({resultsOrErr}) {
-  return (
-    resultsOrErr.constructor.name !== 'Error' &&
-    resultsOrErr.every(
-      r => r.constructor.name !== 'Error' && r.getStatus && r.getStatus() === 'Passed',
-    )
-  );
-}
-
-function updateSpinnerEnd(allTestResults, spinner) {
-  allTestResults.every(didTestPass) ? spinner.succeed() : spinner.fail();
-}
-
-module.exports = updateSpinnerEnd;
